`runner`: do not handle `SIGTSTP` until the process is in the zombies set
Adding processes to ProcessSet is how Runner avoids causing race
conditions where both Ranger and a subprocess could concurrently modify
the terminal settings on Ctrl-Z. But in the time span between when
a process is spawned and when it is added to the ProcessSet, a race
condition is still possible.

To avoid that, Runner will delay handling of the SIGTSTP signal,
"atomically" spawn the process *and* add it to the ProcessSet, and
only after that handle any incoming SIGTSTP signals. This way, only
the subprocess can modify the terminal mode while we're inside the
block on lines 273-275 in `runner.py`, avoiding the race condition.